requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
reflex
reflex==0.8.28.post1